--- services/user/SupervisorSys/ui/src/messaging/index.ts
@@ -1,12 +1,0 @@
-import { IFrameInitialized } from "./IFrameIntialized";
-
-type MessageToParent = IFrameInitialized;
-
-export const sendMessageToParent = (message: MessageToParent) => {
-  window.parent.postMessage(message, "*");
-};
-
-export const sendMessageToService = (service: string) => {
-  // TODO send message
-  console.log(service);
-};
