+++ services/user/SupervisorSys/ui/vite.config.ts
@@ -5,7 +5,7 @@ import { defineConfig } from "vite";
 export default defineConfig({
   resolve: {
     alias: {
-      "@messaging": path.resolve(__dirname, "../../messaging"),
+      "@messaging": path.resolve(__dirname, "../../CommonSys/common/messaging"),
     },
   },
   build: {
